Example refactored
