+++ examples/advanced/customization/strategies/image_preproc.py
@@ -1,0 +1,39 @@
+import warnings
+from typing import Optional
+
+from examples.advanced.customization.implementations.preproc_impls import GammaFiltImplementation
+from fedot.core.data.data import InputData, OutputData
+from fedot.core.operations.evaluation.evaluation_interfaces import EvaluationStrategy
+from fedot.core.operations.operation_parameters import OperationParameters
+from fedot.utilities.random import ImplementationRandomStateHandler
+
+
+class ImagePreprocessingStrategy(EvaluationStrategy):
+    _operations_by_types = {
+        'filter_1': GammaFiltImplementation,
+        'filter_2': GammaFiltImplementation,
+        'filter_3': GammaFiltImplementation
+    }
+
+    def __init__(self, operation_type: str, params: Optional[OperationParameters] = None):
+        self.operation_impl = self._convert_to_operation(operation_type)
+        super().__init__(operation_type, params)
+
+    def fit(self, train_data: InputData):
+
+        warnings.filterwarnings("ignore", category=RuntimeWarning)
+        operation_implementation = self.operation_impl(self.params_for_fit)
+        with ImplementationRandomStateHandler(implementation=operation_implementation):
+            operation_implementation.fit(train_data)
+        return operation_implementation
+
+    def predict(self, trained_operation, predict_data: InputData) -> OutputData:
+        prediction = trained_operation.transform(predict_data)
+        # Convert prediction to output (if it is required)
+        converted = self._convert_to_output(prediction, predict_data)
+        return converted
+
+    def predict_for_fit(self, trained_operation, predict_data: InputData) -> OutputData:
+        prediction = trained_operation.transform_for_fit(predict_data)
+        converted = self._convert_to_output(prediction, predict_data)
+        return converted
